requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
jinja2==3.1.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.29.0